requirements.txt
python-telegram-bot==12.7
numpy==1.18.5
torch==1.5.0
torchvision==0.6.0
opencv-python==4.2.0.34

